feat(main): fix build

+++ src/app/api/projects/[id]/route.ts
@@ -3,22 +3,32 @@ import { NextRequest, NextResponse } from "next/server";
 
 const prisma = new PrismaClient();
 
-export async function GET(
-  req: NextRequest,
-  context: { params: { id: string } }
-) {
-  const { id } = context.params;
+// Ganti code dengan pendekatan ini:
+export async function GET(req: NextRequest) {
+  const urlPath = req.nextUrl.pathname;
+  const id = urlPath.split("/").pop() || ""; // Ambil ID dari URL secara langsung
+
+  // Periksa jika id valid sebelum query ke database
+  if (!id) {
+    return NextResponse.json({ error: "ID is required" }, { status: 400 });
+  }
+
   const project = await prisma.project.findUnique({ where: { id } });
-  return NextResponse.json(project);
+  return project
+    ? NextResponse.json(project)
+    : NextResponse.json({ error: "Project not found" }, { status: 404 });
 }
 
-export async function PUT(
-  req: NextRequest,
-  context: { params: { id: string } }
-) {
-  const { id } = context.params;
+export async function PUT(req: NextRequest) {
+  const urlPath = req.nextUrl.pathname;
+  const id = urlPath.split("/").pop() || "";
   const body = await req.json();
   const { title, description, image, link } = body;
+
+  // Periksa jika id valid
+  if (!id) {
+    return NextResponse.json({ error: "ID is required" }, { status: 400 });
+  }
 
   const project = await prisma.project.update({
     where: { id },
@@ -28,11 +38,15 @@ export async function PUT(
   return NextResponse.json(project);
 }
 
-export async function DELETE(
-  req: NextRequest,
-  context: { params: { id: string } }
-) {
-  const { id } = context.params;
+export async function DELETE(req: NextRequest) {
+  const urlPath = req.nextUrl.pathname;
+  const id = urlPath.split("/").pop() || "";
+
+  // Periksa jika id valid
+  if (!id) {
+    return NextResponse.json({ error: "ID is required" }, { status: 400 });
+  }
+
   await prisma.project.delete({ where: { id } });
   return NextResponse.json({ message: "Project deleted" });
 }
